# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blue Mogul Enterprise Bookkeeping AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export type TransactionSource =
  (typeof TransactionSource)[keyof typeof TransactionSource];

export const TransactionSource = {
  manual: "manual",
  wave: "wave",
  stripe: "stripe",
} as const;

export interface Transaction {
  id: number;
  date: string;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  subcategory?: string | null;
  reconciled: boolean;
  reconciledAt?: string | null;
  source: TransactionSource;
  sourceId?: string | null;
  taxDeductible: boolean;
  notes?: string | null;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface CreateTransactionBody {
  date: string;
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  subcategory?: string;
  taxDeductible?: boolean;
  notes?: string;
}

export type UpdateTransactionBodyType =
  (typeof UpdateTransactionBodyType)[keyof typeof UpdateTransactionBodyType];

export const UpdateTransactionBodyType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export interface UpdateTransactionBody {
  date?: string;
  description?: string;
  amount?: number;
  type?: UpdateTransactionBodyType;
  category?: string;
  subcategory?: string;
  taxDeductible?: boolean;
  notes?: string;
}

export interface ReconcileTransactionBody {
  bankAmount: number;
  notes?: string;
}

export interface Budget {
  id: number;
  year: number;
  month?: number | null;
  category: string;
  amount: number;
  notes?: string | null;
  createdAt: string;
}

export interface CreateBudgetBody {
  year: number;
  month?: number;
  category: string;
  amount: number;
  notes?: string;
}

export interface UpdateBudgetBody {
  amount?: number;
  notes?: string;
}

export interface BudgetSummary {
  category: string;
  budgeted: number;
  actual: number;
  variance: number;
  percentUsed: number;
}

export type TaxItemType = (typeof TaxItemType)[keyof typeof TaxItemType];

export const TaxItemType = {
  sales_tax: "sales_tax",
  estimated_quarterly: "estimated_quarterly",
  payroll: "payroll",
  deduction: "deduction",
} as const;

export interface TaxItem {
  id: number;
  year: number;
  quarter?: number | null;
  type: TaxItemType;
  description: string;
  amount: number;
  dueDate?: string | null;
  paid: boolean;
  paidDate?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateTaxItemBodyType =
  (typeof CreateTaxItemBodyType)[keyof typeof CreateTaxItemBodyType];

export const CreateTaxItemBodyType = {
  sales_tax: "sales_tax",
  estimated_quarterly: "estimated_quarterly",
  payroll: "payroll",
  deduction: "deduction",
} as const;

export interface CreateTaxItemBody {
  year: number;
  quarter?: number;
  type: CreateTaxItemBodyType;
  description: string;
  amount: number;
  dueDate?: string;
  notes?: string;
}

export interface UpdateTaxItemBody {
  description?: string;
  amount?: number;
  dueDate?: string;
  paid?: boolean;
  paidDate?: string;
  notes?: string;
}

export interface QuarterTaxBreakdown {
  quarter: number;
  revenue: number;
  expenses: number;
  netIncome: number;
  estimatedTaxDue: number;
}

export interface DeductionItem {
  category: string;
  amount: number;
}

export interface TaxSummary {
  year: number;
  totalRevenue: number;
  totalExpenses: number;
  netIncome: number;
  totalDeductions: number;
  taxableIncome: number;
  salesTaxCollected: number;
  estimatedTaxPaid: number;
  quarterBreakdown: QuarterTaxBreakdown[];
  topDeductions: DeductionItem[];
}

export interface MonthlyFlow {
  month: string;
  inflow: number;
  outflow: number;
  net: number;
}

export type CategoryFlowType =
  (typeof CategoryFlowType)[keyof typeof CategoryFlowType];

export const CategoryFlowType = {
  income: "income",
  expense: "expense",
} as const;

export interface CategoryFlow {
  category: string;
  amount: number;
  type: CategoryFlowType;
}

export interface CashflowReport {
  startDate: string;
  endDate: string;
  totalInflow: number;
  totalOutflow: number;
  netCashflow: number;
  openingBalance: number;
  closingBalance: number;
  byMonth: MonthlyFlow[];
  byCategory: CategoryFlow[];
}

export interface ProfitLossReport {
  startDate: string;
  endDate: string;
  grossRevenue: number;
  costOfGoodsSold: number;
  grossProfit: number;
  totalExpenses: number;
  netIncome: number;
  incomeByCategory: CategoryFlow[];
  expensesByCategory: CategoryFlow[];
}

export interface DashboardStats {
  currentMonthRevenue: number;
  currentMonthExpenses: number;
  currentMonthNet: number;
  ytdRevenue: number;
  ytdExpenses: number;
  ytdNet: number;
  outstandingTaxes: number;
  unreconciledCount: number;
  cashBalance: number;
  recentTransactions: Transaction[];
  upcomingTaxDates: TaxItem[];
}

export type ReconciliationSessionStatus =
  (typeof ReconciliationSessionStatus)[keyof typeof ReconciliationSessionStatus];

export const ReconciliationSessionStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface ReconciliationSession {
  id: number;
  accountName: string;
  statementDate: string;
  statementBalance: number;
  computedBalance?: number | null;
  difference?: number | null;
  status: ReconciliationSessionStatus;
  completedAt?: string | null;
  createdAt: string;
}

export interface CreateReconciliationSessionBody {
  accountName: string;
  statementDate: string;
  statementBalance: number;
}

export interface SyncResult {
  success: boolean;
  imported: number;
  message: string;
}

export interface IntegrationStatus {
  connected: boolean;
  lastSynced?: string | null;
  message: string;
}

export interface ApiError {
  error: string;
}

export type ListTransactionsParams = {
  type?: ListTransactionsType;
  category?: string;
  startDate?: string;
  endDate?: string;
  reconciled?: boolean;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
  transfer: "transfer",
} as const;

export type ListBudgetsParams = {
  year?: number;
  month?: number;
};

export type GetBudgetSummaryParams = {
  year: number;
  month: number;
};

export type ListTaxItemsParams = {
  year?: number;
  quarter?: number;
  type?: ListTaxItemsType;
};

export type ListTaxItemsType =
  (typeof ListTaxItemsType)[keyof typeof ListTaxItemsType];

export const ListTaxItemsType = {
  sales_tax: "sales_tax",
  estimated_quarterly: "estimated_quarterly",
  payroll: "payroll",
  deduction: "deduction",
} as const;

export type GetTaxSummaryParams = {
  year: number;
};

export type GetCashflowReportParams = {
  startDate: string;
  endDate: string;
};

export type GetProfitLossReportParams = {
  startDate: string;
  endDate: string;
};