# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blue Mogul Enterprise Bookkeeping AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a new budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetBody"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /budgets/summary:
    get:
      operationId: getBudgetSummary
      tags: [budgets]
      summary: Get budget vs actual summary
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetSummary"
...
  /reconciliation/sessions:
    get:
      operationId: listReconciliationSessions
      tags: [reconciliation]
      summary: List bank reconciliation sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReconciliationSession"
    post:
      operationId: createReconciliationSession
      tags: [reconciliation]
      summary: Create a reconciliation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReconciliationSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationSession"
  /reconciliation/sessions/{id}/complete:
    post:
      operationId: completeReconciliationSession
      tags: [reconciliation]
      summary: Complete a reconciliation session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationSession"
...
components:
  schemas:
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        month:
          type: integer
          nullable: true
        category:
          type: string
        amount:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - year
        - category
        - amount
        - createdAt
    CreateBudgetBody:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        category:
          type: string
        amount:
          type: number
        notes:
          type: string
      required:
        - year
        - category
        - amount
    UpdateBudgetBody:
      type: object
      properties:
        amount:
          type: number
        notes:
          type: string
    BudgetSummary:
      type: object
      properties:
        category:
          type: string
        budgeted:
          type: number
        actual:
          type: number
        variance:
          type: number
        percentUsed:
          type: number
      required:
        - category
        - budgeted
        - actual
        - variance
        - percentUsed
...
    ReconciliationSession:
      type: object
      properties:
        id:
          type: integer
        accountName:
          type: string
        statementDate:
          type: string
          format: date
        statementBalance:
          type: number
        computedBalance:
          type: number
          nullable: true
        difference:
          type: number
          nullable: true
        status:
          type: string
          enum: [in_progress, completed]
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - accountName
        - statementDate
        - statementBalance
        - status
        - createdAt
    CreateReconciliationSessionBody:
      type: object
      properties:
        accountName:
          type: string
        statementDate:
          type: string
          format: date
        statementBalance:
          type: number
      required:
        - accountName
        - statementDate
        - statementBalance
...