# lib/api-zod/src/generated/types/createReconciliationSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blue Mogul Enterprise Bookkeeping AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReconciliationSessionBody {
  accountName: string;
  statementDate: Date;
  statementBalance: number;
}