# lib/api-zod/src/generated/types/reconciliationSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blue Mogul Enterprise Bookkeeping AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ReconciliationSessionStatus } from "./reconciliationSessionStatus";

export interface ReconciliationSession {
  id: number;
  accountName: string;
  statementDate: Date;
  statementBalance: number;
  computedBalance?: number | null;
  difference?: number | null;
  status: ReconciliationSessionStatus;
  completedAt?: Date | null;
  createdAt: Date;
}